requirements.txt
streamlit
numpy
Pillow
python-dotenv
diffusers
transformers
torch
accelerate
SpeechRecognition
pydub
sounddevice
soundfile
requests